Shutdown gracefully on unexpected server exception (#6712)
## What is the goal of this PR?

To prevent zombie or broken server states due to unexpected exceptions, we no longer catch and swallow unexpected exceptions. Instead, we shutdown the server. The catch-and-swallow behaviour is dangerous because 1) an uncaught exception means a thread is also killed somewhere, and could put the server into a zombie or otherwise broken state, and specifically can cause 2) zombie state by catching an OOM exception which would stall the server instead of letting it die. 

## What are the changes implemented in this PR?

* Allow TypeDB to exit if there's an unexpected exception by removing the uncaught exception handler