Implement fast termination for reasoner (#6794)
## What is the goal of this PR?

We force each transaction-bound reasoner instance to terminate rapidly when any exception occurs, such as those generated when the transaction is closed while queries are still executing. This fixes a bug where queued up reasoning work would continue running for some time even after transaction close. Previously, scenarios where many queries ran concurrently and over a longer period of time, the incurred resource consumption could add up to bottleneck the entire server.

## What are the changes implemented in this PR?

* Introduce preemption (eg. skip the message queue) in the Actor model, by switching the internal actor queue to be a `BlockingDeque` instead of a `TransferQueue`
* Introduce a new `terminate` API on actors, which should generally be used with the new `Actor.Driver.executeNext()` execution mechanism.
* Tasks submitted to the actor executors automatically check whether they should execute, by checking if the actor that submitted them is terminated or not, allowing the messages in the queue to drain quickly.

### Performance implications:
We did a a few synthetic benchmarks to check whether the replacement of the `LinkedTransferQueue` is much slower than the `LinkedBlockingDeque` and found that:
1. Under no contention (single writer/single reader) the `LinkedTransferQueue` can be faster (variance is higher, but between -10 and +30% faster) compared to the `LinkedTransferDeque`
2. Under any contention, the `Deque` and the `Queue` are actually very comparable, and the `Deque` is sometimes faster.

These tradeoffs are acceptable for our use cases so we proceed with the `Deque` that provides the ability to insert at the start of the queue natively.