Relax rule validation to allow unanswerable but coherent rules (#6801)
## What is the goal of this PR?

We allow rules to be written in a way that is semantically sensible in the schema, but may not produce any answers (specifically, if `when` types can't have any instances because they are abstract). 

We therefore match the current behaviour of `match` queries: `match $r isa relation;` does not throw any exceptions even if there are no concrete subtypes of `relation`. As a result, after this change, it will be possible to write rules that will never trigger any inferences, since never match any concrete types.

This change allows writing and sharing 'general' rules that work over an abstract schema, only to be specialised at a later part of the development cycle.

Note that a rule `then` must still be written using concrete, insertable types.

## What are the changes implemented in this PR?

We relax rule validation to log a server-side warning if the rule is not answerable, but it is coherent (eg. it makes sense in the schema).