Fixed variable in insert "out of scope of match query" in update query (#6823)
## What is the goal of this PR?

We fix the filtering of variables in insert clause of update queries to allow inserting new entities.

Fixes #6549 

## What are the changes implemented in this PR?

For the output of the `match` query that subsequently feeds into delete and insert, the filters of the `get` are the intersection of the `insert` and the `match` query, as well as the variables in the `delete`. Further, the set of variables in the `delete` query should be a subset of the set of variables in `match` query.