Conjunction stream graph optimisation (#6826)
## What is the goal of this PR?

Enable the reasoner to efficiently cache and reuse results to
sub-conjunctions where possible. This can dramatically increasing the
performance of reasoning in some large conjunctions configurations.

## What are the changes implemented in this PR?

* Reuse results to sub conjunction wherever possible
- The current architecture requires us to buffer the results if we have
multiple readers to the stream.
* Aggressively prune out variables which are not relevant to the
remaining resolvables in the plan
- If the variable is an output of the conjunction, we extend the result
of the sub conjunction with it.
  - else we just project it out.
* Introduce ConjunctionStreamPlan, which divides the conjunction in a
way that maximises reusability.

### Motivation
The stream plan forms a tree, with ResolvablePlans as the leaves, and
CompoundStreamPlan as internal nodes. Each CompoundStreamPlan has a left
and right child. Each node in the tree represents some sub-conjunction.
The order of the leaves from left-to-right matches the ordering in the
plan chosen by ReasonerPlanner.

The actual stream graph forms a tree, with the individual Resolvables as
the leaves and compound streams as internal nodes. Each node is
identified by the bounds to a set of variables - henceforth
'**identifiers**'. A compound stream has a unique left child, and a
family of right-children. Each right-child corresponds to the same sub
conjunction but differs on the identifiers. The changes correspond to
normalising intermediate tables if we saw everything as tables.
  
**Current system**
In the current system, Every conjunction is divided into its first
resolvable on the left, and the remaining resolvable on the right. We
create a new compound stream for each branch in the conjunction, whose
identifiers are the current bounds for all variables in the resolvables
processed so far.

For the goal : `z(A,D) :- p(A, B), q(B, C), r(C, D).` (with no bounds to
Z) this would correspond to the plan:

```
⨂: pqr(A,B,C,D)
+---?: p(A,B)
^---⨂: qr(a, b,C,D)
    +---?: q(b, C)
    ^---?: r(c,D)
```
Where ,  
* ⨂ represents the a compound stream node
* ? represents a resolvable node
* `q(b,C)` indicates it only contains the rows in `q` which are queried
by backward-chaining (with values for `b` resulting from the resolvables
before it, here {`p`}),
* `qr(a,b,C,D)` indicates the intermediate result of joining `q` with
`r`, again with only results `a,b` from p.

The plan produces the stream graph:
**The amount of work done is approximately number of answers propagated
up each edge**

![image](https://github.com/vaticle/typedb/assets/4918770/e8663894-d571-4218-9e55-c093676ffdd5)


**Reusing results**
Considering that we only require {A,D} as output, we transform the query
after each join - projecting away the variables not used in the suffix,
and joining any outputs on the way back up. Further, it's clear that the
value of `qr`does not actually depend on the value of `a`, so could be
reused by different values of a which share the same value for `b`.

This results in the plan:
```
⨂: pqr(A,D)
+---?: p'(A,B)
^---⨁: (a,)
    +---⨂: qr(b, D)          (reusable)
        +---?: q'(b, C)
        ^---?: r'(c,D)
```

Where, 
* ⨁" (a,) is the newly introduced map stream which appends the
**extension** bounds (the values of `a`) to the results from the
subtree.

This would be the resulting stream graph. Notice we also de-duplicate
answers this way.

![image](https://github.com/vaticle/typedb/assets/4918770/35ad7bdf-767e-494c-b029-84b297927f0f)

** Optimising reusability **
Suppose our goal was instead, `z(A,D) :- p(A, B), q(B, C), r(C, D),
s(A,D).` [^1]

We would not be able to reuse results for the suffix `qrs` because of
the dependency on `a`. i.e. `qrs(a,b,D)` and `rs(a,c,D)` would be the
sub conjunctions.
```
⨂: pqrs(A,D)
+---?: p'(A,B)
+---⨂: qrs(a, b, D)
       +---?: q'(b, C)
       ^---?: rs'(a,c,D)             (reusable)
               +---?: r(c, D)
               ^---?: s(a,d)
```
with the stream graph:

![image](https://github.com/vaticle/typedb/assets/4918770/2a2be0dd-5b2b-48c0-9222-ae1fad483b55)


There may be multiple ways to restructure the stream, but we aim to pick
a structure such that the bounds of every right-child is bound with
exactly the set of variables expected by its left-child - The only
violation allowed is for a sub conjunction of size 2, where it's
unavoidable.

```
⨂: pqrs(A,D)
+---⨂: pqr'(A,B)             
|   +---?: p'(A,B)
|   ^---⨁: (a,)
|       ^---⨂: qr(b, D)             (reusable)
|           +---?: q'(b, C)
|           ^---?: r'(c,D)
^---?: s(a,d)
```
with the stream graph:

![image](https://github.com/vaticle/typedb/assets/4918770/ecba0675-e4d7-4b20-ba3b-4f134728d8d2)


##### Flattening  
In cases where a compound stream node is redundant, we directly add it's
children to it parent, reducing one hop.
This is the case when
a) The parent is the exclusive reader AND
b) No bounds for the compound stream are bound by the parent's other
(more-left) children.

#### Implementation
##### reusing answers
When we create a compound stream for a conjunction
1. Create a for the left child with the relevant subset of bounds as
identifier
  * The left child will 'stream' results in as and when it computes one.

For each answer received back from its left child, We create (or reuse)
a right-child by:
1. Create (or get an existing) compound stream with the relevant
identifier bounds.
2. Compute a set of extension bounds - consisting of the variables bound
by the left-child, not used by the right BUT is either used by a later
resolvable or an output of the conjunction.
- If this set is non-empty, we attach the compound stream through a
MapStream which extends any answers coming from it with the extension
bounds.
  - else, we attache the compound stream from one directly.
* The right child will stream results in as and when it computes them
(through the map stream)

We propagate any answer coming from our right children up towards the
parent compound-stream or processor.

---------

Co-authored-by: joshua <joshua@vaticle.com>