Add RPC logging for session and transaction creation (#6836)
## What is the goal of this PR?

To monitor production instances of TypeDB and TypeDB enterprise more effectively, when using debug logging the server logs open client connections (each sessions and number of transactions each) at a rate of once per minute. When there are no connections, the server does not log to avoid noise in the idle state.

To enable these logs, set the logger for 'com.vaticle.typedb.core.server` to DEBUG in the TypeDB configuration file.

## What are the changes implemented in this PR?

* Add periodic logging of the global connection state when the server is in DEBUG mode

### Design

We considered alternatives:
1. Not logging the global state each time -- this will not be useful for debugging and monitoring
2. Logging global connection state on client connection -- this makes sense, but probably isn't often enough to be useful
3. Logging at each connection/session/transaction  -- this makes sense, and is consistent, but can get expensive. It also requires transaction open/close to query the _other_ sessions to log their information, which is an awkward control flow

In the end, the best tradeoff of usefulness, performance, usability, and 'correct' control flow by domain was decided to lie in a polling loop that lives at the top-level service.

Note that we ultimately want a client RPC method to query the current server state, which can retrieve network state (sessions/transactions) as well as storage statistics, etc.