Fix server hang-ups during shutdown (#6842)
## What is the goal of this PR?

Under exceptional circumstances, such as when the server runs out of memory, the server could fail to shut down. We modify the server shutdown process and unexpected exception handler to shutdown in several stages, in the most severe case halting the JVM runtime immediately.

## What are the changes implemented in this PR?

Implement graceful and forceful shutdown from within the process:

1. If an unexpected exception occurs, we try to perform the server `close()`. Only 1 thread is allowed to do this one time, and if it fails to do so it throws a checked exception. The caller of `close()` is expected to handle this case and force a halt of the JVM
2. If the user hits ctrl+c, we add a shutdown hook to gracefully close the server, but do nothing if it fails to since the JVM is already shutting down

During the server `close()`, we also now force the grpc server layer to shut down in 10 seconds.