Deletes are idempotent instead of throwing (#6884)
## What is the goal of this PR?

We move partially towards the 'assert-style' deletion outlined in #6882
to allow idempotent deletion. Previously, deletes would throw when
trying to delete a previously-removed/non-existent connection.

We now guarantee that all deletes will continue even if trying to delete
a relationship, ownership, role player, or concept that doesn't exist
(ie. was previously deleted).

For example, the following is now valid:
```
match 
$x has $a, has $b; $b = $a;
delete
$x has $a;
```
Because the same attribute would be generated twice by the `match`, the
second delete would fail and throw an error. This is now an idempotent
operation, which will continue without an error.

We also introduce a new operation to disallow a new class of
semantically unsafe queries. We add an explicit 'runtime' type check
before performing the idempotent delete operation. This means the
following semantically invalid query will still throw an exception when
any data matches:
```
match
$x isa person;
$id isa company-id;
delete
$x has $id;
```

## What are the changes implemented in this PR?

* Relax delete semantics to allow idempotent deletion
* Add a 'runtime' type check that before trying to perform any
delete-if-present operation, can throw a meaningful error if a
connection being deleted is not allowed in the schema at all.