Ignore data directories that aren't TypeDB databases (#6938)
## Release notes: usage and product changes

TypeDB ignores directories in its configured 'data' directory that do
not contain the subdirectories 'data' and 'schema'. In the past, the any
directory was loaded as a TypeDB database, which could cause the server
to crash on bootup.

This should help with OS-created directories in some system (such as
Lost+Found), or users' debugging directories.

## Implementation

* Instead of validating and crashing when the data directory does not
conform to the expected shape, we now ignore the directory and proceed.