Update entry point runner for better UX (#6951)
## Usage and product changes

We update to the updated entry point for the assembled TypeDB Core,
which omits `console` from the usage help if it is not present to avoid
confusion. See https://github.com/vaticle/typedb/issues/6942 for more
details.

The command to boot up the server (unchanged):
```
typedb server --server.address=<address>
``` 
The command to boot up the console:
```
typedb console --core=<address>
```

We also improve the UX of the windows version of the entry point.
Console no longer opens in a new window, `--help` is printed inline as
expected, and in the event of the server failure, the logs are displayed
to the user. See https://github.com/vaticle/typedb-common/pull/158 for
more details.