Fix negated traversals so filters apply after the negation (#6967)
## Usage and product changes
Fixes a bug where the `get` clause filtered out concepts before the
negation was applied. This lead to the negation being evaluated without
that variable being bound.

## Implementation
* Refactor code so the `Modifier.Filter` is not passed to the traversal
if a negation is present. Instead the answers coming out of the negation
are filtered afterwards using `FunctionalIterator.filter`. This matches
an existing case which was correctly implemented.