Sync dependencies in CI (#6971)
## Usage and product changes

We add a sync-dependencies job to be run in CI after successful snapshot
and release deployments. The job sends a request to vaticle-bot to
update all downstream dependencies.

Note: this PR does _not_ update the `dependencies` repo dependency. It
will be updated automatically by the bot during its first pass.