Generate persistent server ID (#6978)
## Usage and product changes

We update the server ID to be generated once, then persisted into the
server's data directory. A fallback ID of of `_0` is also used if this
fails for any reason.

## Implementation

Generate a length 16 string of upperchase ASCII character and digits,
persisted into the data directory. On bootup we read this value, if it
exists.