Allow Type IDs of deleted types to be reused (#7040)
## Usage and product changes
Changes how type IDs are allocated so Type IDs from deleted types can be reused and allocated to newly created types. 
If we delete a type, we do not reset the corresponding thing ID generator to 0 - because it's non-trivial and we don't have to since U64 will have plenty of free values. Hence, If a type was created with a re-used TypeID, its Object ID of its instances will not start from 0.

## Implementation
* Instead of just incrementing an AtomicU16, we iterate over the `zip(range_of_type_ids, type_ids_in_db)` ids and return the first "gap" ( `it.0 != it.1` )
* We store an offset to start the next iteration from, so that we do a maximum of two seeks for allocating multiple Type IDs.
* If we don't find a free TypeID from our offset, we do another scan from the beginning (offset  = 0)