Patterns (#7078)
## Usage and product changes

We introduce the `//ir` package, which contains the initial design for
the new intermediate-representation (IR) for patterns and functions. We
also introduce the idea of intermediate representation of `Program`,
which will hold the entry point pattern and all required functions -
these can all have type inference and optimisation passes applied
together.

Variable categorisation, type inference, and "compiler" optimisation
passes will live in the `//ir` package.

We also place packages and mostly empty modules for `//query`,
`//answer` and `//function`.
We aim to have the following dependency graph:

```
Query -> Function, IR, Answer, (Executor - not introduced yet), Concept
Function -> IR (maybe Answer?)
IR -> Concept
(Executor -> IR, Answer, Concept)
```

In other words, a Query Manager will create an IR `Program` for the
query and invoked functions, and apply type inference and any 'compiler'
optimisation passes. This IR will then be handed over to
traversal/executor to query plan, and execute (Query plan + IR).

### IR design

The IR is build around the idea of a Functional Blocks. The `Program` is
a functional block, as well as each `FunctionIR` it conatins. A
Functional Block consists of a Pattern along with any Modifiers.

A Pattern is a destructured IR of TypeQL. We will aim to break down all
operations into their simplest form:
```
$x has name "john"
->
$x has $1
$1 isa $2
$2 type name
$1 == "john"
```
or
```
$r ($x, $y, friend: $z) isa friendship;
->
$r isa $t
$t type friendship
$r roleplayer $x
$r roleplayer $y
$r roleplayer $z (role: $1)
$1 sub friendship:friend
```

We also remove all names from the IR, using a numbering scheme for
variables, and keeping a mapping from variable to Name in a shared
PatternContext. PatternContext is a hidden construct, but helps with
tracking variable declarations, validation, etc.