Add operation time validations for schema. Finish inheritance, overrides, ordering (#7094)
## Usage and product changes
We implement ConceptAPI schema operation time validation for all the
main features of TypeDB:
* Differentiate `put` and `create` operations for data creation: `put`
is ignorant if the *thing* already exists in the db, while `create`
returns a specific error if it already exists.
* Validate all the `sub` and `override` connections for each related
action.
* Be specific while working with declared properties and interfaces or
both declared and inherited ones.
* Make sure that annotations are set only for correct types and
interfaces, and that they do not conflict with each other on one type or
through inheritance.
* Make sure that ordering is stable for the whole inheritance line.
* While checking inheritance issues, try to ban (non-blocking) the
maximum of invalid operations. The current idea is to ban
incompatibilities with supertypes, leaving subtypes checks for commit
time.

All these validations aim to fulfill the BDD tests declared in
https://github.com/vaticle/typedb-behaviour/pull/287, all the needed BBD
steps implementations are a part of this PR as well.

## Implementation

We try to follow this flow of operations:
* When receiving a call for a concept api interface, call a
`type_manager` function with a minimum of additional operations possible
* `type_manager` has two stages of processing such requests: validation
and execution. We use `operation_time_validation` for operation time
validations, and return errors if these validations are not successful.
Here, we try to balance between "get as much data to pass to validation
as possible" and "let validation decide what additional data to read
from the db by itself" based on common sense, both of the approaches are
comfortable in specific cases.
* `operation_time_validation` performs validation and defines Errors. In
most cases, these validations are targeted to one specific feature, but
sometimes it's handy to have a general method of validations to return
one of multiple possible errors of one "subset" of errors.

All the main properties and edges have their caches with
declared/transitive instances (transitive does not have a prefix/postfix
anymore for interface simplicity, it should be the main point of getting
properties and interface implementations):
* declared *things* getters return just these *things*.
* non-declared *things* getters return maps of {*thing*, *source*},
where *source* is the supertype/overridden edge where this *thing* is
inherited from.

All the other pieces of the code try to follow the basics put there in
the branch before, just filling the gaps.