Schema commit time validations (#7104)
## Usage and product changes
We add schema commit time validations to ensure that the schema is fully
correct after a series of operations. These checks mostly concentrate on
the things that are ignored in the operation time validations (to let
users modify their schemas in an easier manner), like:
* verify that no types/edges conflict with their supertypes/overridden
edges declarations (we usually don't check subtypes when we change a
type for operation time validations, with minor exceptions) e.g.
annotations and ordering
* verify that capabilities are still correctly overridden after a series
of subtyping moves (e.g. we can only override the supertype interface
type)
* verify that cardinality is correct for all levels of inheritance ([see
details](https://discord.com/channels/665254494820368395/699946888572567611/1262347400232828929))

We add new annotations: `@values` and `@range`, and refactor the usage
of cardinality on the `concept` level.
Now, when you call for `get_cardinality` of a `Capabiity` (`owns`,
`plays`, `relates`), you get the capability's constraint, which is
either an explicitly set annotation or a default value. Thus, all
capabilities have `cardinality` constraints, and these constraints can
be modified through setting `@card` annotations.

We also change parsing of annotations to align with the design (`x..y`
instead of `x, y`). Additionally, most of the checks for `fails` in
tests now don't accept `ConceptReadErrors` so as not to miss typos and
serious inner issues of the system!

## Implementation

All the commit time validations are written in `commit_time_validation`
file and are triggered by `CommitTimeValidation::validate` function.

The algorithm of the type validation is the following:
1. Get root type (entity/relation/attribute)
2. Validate this type and all its declared capabilities (`owns`,
`plays`, `relates`)
3. Repeat step 2 for all the transitive subtypes of the root (all the
existing types)

We also additionally validate all existing `struct`s.

Common validation functions used in both `commit_time_validation` and
`operation_time_validation` are not in `validation.rs`.

Refactoring:
* `InterfaceImplementation` became `Capability` to describe `Owns`,
`Plays`, and `Relates`. The previous term didn't quite align with
`Relates`
* `get_annotations` and `get_cardinality` are now declared in
`Capability`. `set_annotation`s is still just a function for specific
types. It allows us to shrink some code (especially that checks `card`s)
to one function instead of 3 duplicates
* we introduce `CapabilityKind` so it could be used in places like
`Errors` (for the sake of code shrinking as well)
* we shrink some of the common annotation setting/unsetting validations
in `type_manager` so nothing is lost!